+++ RELEASES.md
@@ -1799,3 +1799,13 @@ Based on:
 - [typescript v0.9.0-alpha.15] .
 ### Releases
 - [NPM v0.9.0-alpha.15] https://www.npmjs.com/package/cribl-control-plane/v/0.9.0-alpha.15 - .
+
+## 2026-06-13 03:45:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.776.0 (2.900.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.9.0-alpha.16] .
+### Releases
+- [NPM v0.9.0-alpha.16] https://www.npmjs.com/package/cribl-control-plane/v/0.9.0-alpha.16 - .